fixed token response model
