+++ src/Obsidian/Controllers/ApiControllers/AccountController.cs
@@ -1,16 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Claims;
-using System.Threading.Tasks;
 
 namespace Obsidian.Controllers.ApiControllers
 {
-
     [Route("api/[controller]/[action]")]
     public class AccountController : Controller
     {
-        
     }
 }
