+++ src/Obsidian/Controllers/OAuth/OAuth20Controller.cs
@@ -1,20 +1,13 @@
 ﻿using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Caching.Memory;
 using Obsidian.Application.OAuth20;
 using Obsidian.Application.ProcessManagement;
 using Obsidian.Domain;
-using Obsidian.Domain.Repositories;
 using Obsidian.Models;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-
-//TODO: remove this when implemented
-#pragma warning disable CS1998
-
 
 namespace Obsidian.Controllers.OAuth
 {
@@ -27,37 +20,6 @@ namespace Obsidian.Controllers.OAuth
         {
             _dataProtector = dataProtectionProvicer.CreateProtector("Obsidian.OAuth.Context.Key");
             _sagaBus = bus;
-        }
-
-        [Route("oauth20/authorize/frontend/signin")]
-        [HttpGet]
-        public IActionResult FrontEndSignInDebug()
-        {
-            return View("SignIn");
-        }
-
-
-        [Route("oauth20/authorize/frontend/grant")]
-        [HttpGet]
-        public IActionResult FrontEndGrantDebug()
-        {
-            var context = _dataProtector.Protect(Guid.NewGuid().ToString());
-            var client = Client.Create(Guid.NewGuid(), "http://za-pt.org/exchange");
-            client.DisplayName = "iTech";
-            ViewBag.Client = client;
-            ViewBag.Scopes = new[] {
-                PermissionScope.Create(Guid.NewGuid(),"obsidian.basicinfo","Basic Information","Includes you name and gender."),
-                PermissionScope.Create(Guid.NewGuid(),"obsidian.email","Email address","Your email address."),
-                PermissionScope.Create(Guid.NewGuid(),"obsidian.admin","Admin access","Manage the system."),
-            };
-            return View("PermissionGrant", new PermissionGrantModel { ProtectedOAuthContext = context });
-        }
-
-        [Route("oauth20/authorize/frontend/denied")]
-        [HttpGet]
-        public IActionResult FrontEndDeniedDebug()
-        {
-            return View("UserDenied");
         }
 
         [Route("oauth20/authorize")]
@@ -93,55 +55,23 @@ namespace Obsidian.Controllers.OAuth
             {
                 case OAuth20State.RequireSignIn:
                     return View("SignIn", new OAuthSignInModel { ProtectedOAuthContext = context });
+
                 case OAuth20State.RequirePermissionGrant:
                     return PermissionGrantView(result);
+
                 case OAuth20State.AuthorizationCodeGenerated:
                     return AuthorizationCodeRedirect(result);
+
                 case OAuth20State.Finished:
                     return ImplictRedirect(result);
+
                 default:
                     return BadRequest();
             }
         }
-
-        private IActionResult PermissionGrantView(OAuth20Result result)
-        {
-            var context = _dataProtector.Protect(result.SagaId.ToString());
-            ViewBag.Client = result.PermissionGrant.Client;
-            ViewBag.Scopes = result.PermissionGrant.Scopes;
-            return View("PermissionGrant", new PermissionGrantModel { ProtectedOAuthContext = context });
-        }
-
-        private IActionResult ImplictRedirect(OAuth20Result result)
-        {
-            var tokenRedirectUri = BuildImplictReturnUri(result);
-            return Redirect(tokenRedirectUri);
-        }
-
-        private IActionResult AuthorizationCodeRedirect(OAuth20Result result)
-        {
-            var codeRedirectUri = $"{result.RedirectUri}?code={result.AuthorizationCode}";
-            return Redirect(codeRedirectUri);
-        }
-
-        private static string BuildImplictReturnUri(OAuth20Result result)
-        {
-            var sb = new StringBuilder($"{result.RedirectUri}?access_token={result.Token.AccessToken}");
-            if (result.Token.AuthrneticationToken != null)
-            {
-                sb.Append($"&authentication_token={result.Token.AuthrneticationToken}");
-            }
-            if (result.Token.RefreshToken != null)
-            {
-                sb.Append($"&refresh_token={result.Token.RefreshToken}");
-            }
-            var tokenRedirectUri = sb.ToString();
-            return tokenRedirectUri;
-        }
-
         [Route("oauth20/authorize")]
         [HttpPost]
-        public async Task<IActionResult> Authorize([FromForm]OAuthSignInModel model)
+        public async Task<IActionResult> SignIn([FromForm]OAuthSignInModel model)
         {
             Guid sagaId;
             var context = _dataProtector.Unprotect(model.ProtectedOAuthContext);
@@ -161,12 +91,16 @@ namespace Obsidian.Controllers.OAuth
                 case OAuth20State.RequireSignIn:
                     ModelState.AddModelError(string.Empty, "Singin failed");
                     return View("SignIn");
+
                 case OAuth20State.RequirePermissionGrant:
                     return PermissionGrantView(result);
+
                 case OAuth20State.AuthorizationCodeGenerated:
                     return AuthorizationCodeRedirect(result);
+
                 case OAuth20State.Finished:
                     return ImplictRedirect(result);
+
                 default:
                     return BadRequest();
             }
@@ -188,10 +122,13 @@ namespace Obsidian.Controllers.OAuth
             {
                 case OAuth20State.AuthorizationCodeGenerated:
                     return AuthorizationCodeRedirect(result);
+
                 case OAuth20State.Finished:
                     return ImplictRedirect(result);
+
                 case OAuth20State.UserDenied:
                     return View("UserDenied");
+
                 default:
                     return BadRequest();
             }
@@ -206,13 +143,15 @@ namespace Obsidian.Controllers.OAuth
                 var message = new AccessTokenRequestMessage(model.Code)
                 {
                     ClientId = model.ClientId,
-                    ClientSecret = model.ClientSecret
+                    ClientSecret = model.ClientSecret,
+                    Code = model.Code
                 };
                 var result = await _sagaBus.SendAsync<AccessTokenRequestMessage, OAuth20Result>(message);
                 switch (result.State)
                 {
                     case OAuth20State.AuthorizationCodeGenerated:
                         return BadRequest();
+
                     case OAuth20State.Finished:
                         return Ok(new AccessTokenResponseModel
                         {
@@ -227,5 +166,78 @@ namespace Obsidian.Controllers.OAuth
             }
             return BadRequest();
         }
+
+        private static string BuildImplictReturnUri(OAuth20Result result)
+        {
+            var sb = new StringBuilder($"{result.RedirectUri}?access_token={result.Token.AccessToken}");
+            if (result.Token.AuthrneticationToken != null)
+            {
+                sb.Append($"&authentication_token={result.Token.AuthrneticationToken}");
+            }
+            if (result.Token.RefreshToken != null)
+            {
+                sb.Append($"&refresh_token={result.Token.RefreshToken}");
+            }
+            var tokenRedirectUri = sb.ToString();
+            return tokenRedirectUri;
+        }
+
+        #region Results
+
+        private IActionResult PermissionGrantView(OAuth20Result result)
+        {
+            var context = _dataProtector.Protect(result.SagaId.ToString());
+            ViewBag.Client = result.PermissionGrant.Client;
+            ViewBag.Scopes = result.PermissionGrant.Scopes;
+            return View("PermissionGrant", new PermissionGrantModel { ProtectedOAuthContext = context });
+        }
+
+        private IActionResult ImplictRedirect(OAuth20Result result)
+        {
+            var tokenRedirectUri = BuildImplictReturnUri(result);
+            return Redirect(tokenRedirectUri);
+        }
+
+        private IActionResult AuthorizationCodeRedirect(OAuth20Result result)
+        {
+            var codeRedirectUri = $"{result.RedirectUri}?code={result.AuthorizationCode}";
+            return Redirect(codeRedirectUri);
+        }
+
+        #endregion Results
+
+        #region Front-end debug
+
+        [Route("oauth20/authorize/frontend/signin")]
+        [HttpGet]
+        public IActionResult FrontEndSignInDebug()
+        {
+            return View("SignIn");
+        }
+
+        [Route("oauth20/authorize/frontend/grant")]
+        [HttpGet]
+        public IActionResult FrontEndGrantDebug()
+        {
+            var context = _dataProtector.Protect(Guid.NewGuid().ToString());
+            var client = Client.Create(Guid.NewGuid(), "http://za-pt.org/exchange");
+            client.DisplayName = "iTech";
+            ViewBag.Client = client;
+            ViewBag.Scopes = new[] {
+                PermissionScope.Create(Guid.NewGuid(),"obsidian.basicinfo","Basic Information","Includes you name and gender."),
+                PermissionScope.Create(Guid.NewGuid(),"obsidian.email","Email address","Your email address."),
+                PermissionScope.Create(Guid.NewGuid(),"obsidian.admin","Admin access","Manage the system."),
+            };
+            return View("PermissionGrant", new PermissionGrantModel { ProtectedOAuthContext = context });
+        }
+
+        [Route("oauth20/authorize/frontend/denied")]
+        [HttpGet]
+        public IActionResult FrontEndDeniedDebug()
+        {
+            return View("UserDenied");
+        }
+
+        #endregion Front-end debug
     }
 }
